Se agregan archivos (Recaudo) a TramiteBundle
Se agregan las entidades Recaudo y TipoRecaudo a TramiteBundle y se
elimina Tipo_tramite.php repetido
